+++ NEWS.md
@@ -19,7 +19,7 @@
 * Function `summary.fit_model_b()` returns now the model summary (in addition to printing to the console).
 * Function `robpredict()` 
   * The function gained the additional argument `seed`  to specify the random seed in the parametric bootstrap.
-  * The progress bar in function `robpredict()` is now implemented with `utils:txtProgressBar()` and can be switched off using argument `progress_bar`.
+  * The progress bar in function `robpredict()` is now implemented with `utils:txtProgressBar()` and can be switched off using  `progress_bar = FALSE`.
 
 
 ## MISC